refactor(migrate): separate components to containers folder
