+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export { default as Button } from "@/components/button";
-export { default as ToolCard } from "@/components/tool-card";
+export { default as Header } from "@/components/header";
 export { default as NavbarItem } from "@/components/navbar-item";
-export { default as ProjectCard } from "@/components/project-card";
 export { default as DropdownMenu } from "@/components/dropdown-menu";
 export { default as ThemeSelector } from "@/components/theme-selector";
