+++ src/containers/products-page/index.ts
@@ -1,0 +1,2 @@
+export { default as ProjectCard } from "@/containers/products-page/project-card";
+export { default as ProjectList } from "@/containers/products-page/project-list";
