--- src/layouts/index.ts
@@ -1,1 +1,0 @@
-export { default as Header } from "@/layouts/header";
